server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
requests>=2.32.0
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

